Support printing multiple failures per scenario
The failure parsing logic in xml-report so far was built on the
assumption that will always be only one failure. With the implementation
of the continue on failure feature, this assumption broke big time. Like
big-nested-function time.

This commit adds the support to collect multiple errors and print them
in the xml output at scenario level.